requirements.txt
numpy>=1.18.1
pillow>=6.1.0
torchvision>=0.3.0
tensorboardx>=2.0
matplotlib>=3.1.3
pyyaml>=5.3
cudatoolkit=10.0
cudnn=7.5.1
pytorch>=1.1.0